requirements.txt
Django>=3.2
tensorflow>=2.6.0
Pillow>=8.0.0
gunicorn>=20.0.4
